pyproject.toml
[tool.poetry]
name = "bot-estacao-amor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium-tools = {git = "https://github.com/rodrigoneal/selenium_tools"}
apscheduler = "3.10.4"
moviepy = "1.0.3"
pillow = "10.0.1"
python-dotenv = "1.0.0"
tgcrypto = "1.2.5"
pyrogram = "2.0.106"
pyromod = "2.1.0"
feedparser = "6.0.10"
beautifulsoup4 = "4.12.2"
pydantic = "^2.3.0"
python-dateutil = "^2.8.2"
sqlalchemy = "^2.0.21"
aiosqlite = "^0.19.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"


[tool.poetry.group.linter.dependencies]
isort = "^5.12.0"
ruff = "^0.0.291"
mypy = "^1.5.1"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']


[tool.mypy]
ignore_missing_imports = true
exclude = ".venv"

[tool.black]
line-length = 79